Worked on smoothing plot, updated paths for OSX, found bug in theta()
Fixed bug in smoothing evaluation because a was being raised to a power
instead of 2.

If running binary from project root, then paths for OSX do not need
to cd back a directory. (if running from bin/ then it will)

In theta(), the clever evaluation without conditionals allows for
nan when x = 0.0 (even for functions that are well defined).