Split theta into theta and theta_star, fixed bug in smoothing calculation involved a, adjusted y-axis scaling on plot, etc
Instead of trying to fix the clever condition-less code in previous theta, it seemed to be easier and make more sense to just split the function into the two functions that it represented.

In smoothing calculation, a1 became al (a-one, a-ell), for each level a1 = a1/2 instead of a1 = 1/(2a).

In the plot, the y-axis was [-1.0,10] and is now [0.0,X] where X = 1.33*g_1(0.0).

There may also be some minor changes to comments.