Moved all of the driver-type code into gamma/phi
The initial driver code which tested correctness has moved from main
to the respective files.

Additionally, the code to create Figure 1 in the paper is in gamma

Lastly, main is now a while-loop with a menu system to allow the user
to select which function they want to display or test.
ve=i