Added asserts to gamma&phi, started nested phi plot.
Some dynamic memory allocates were not asserted and so I've added them
where I noticed in gamma and phi code.

Main was modified by adding two new entries for Figures 4 and 5 in the
paper. These will be the nested B-splines and semi-nested C1 functions.

NOTE: There appears to be a problem with the current nested B-spline
calculation b/c the sum of the fine grid values do not equal the course
grid value.