Fixed sum of fine grids, wrote code for command file.
The fine grid values needed to be multiplied by the coefficients which
were calculated in a previous commit. This fixed the sum of the fine grids
matching the coarse grid.

I played with different gnuplot options in the command file. It seems that
dashed lines on the windows terminal is not easy?